Update Go base image in pre-processor Dockerfile
This commit updates the base image in the `pre-processor` Dockerfile from `golang:1.24.3-alpine` to `golang:1.25-alpine`. This change ensures that the service benefits from the latest features, performance improvements, and security updates available in the newer Go version.

Key changes:
- Updated base image version in `Dockerfile.preprocess`.

This update aims to maintain the service's reliability and performance by leveraging the latest Go enhancements.